Class body ending with a semicolon throws a SyntaxError https://bugs.webkit.org/show_bug.cgi?id=144244
Reviewed by Darin Adler.

Source/JavaScriptCore:

The bug was caused by parseClass's inner loop for method definitions not moving onto the next iteration
it encounters a semicolon. As a result, we always expected a method to appear after a semicolon. Fixed
it by continue'ing when it encounters a semicolon.

* parser/Parser.cpp:
(JSC::Parser<LexerType>::parseClass):

LayoutTests:

Added a regression test for having a semicolon inside the class definition.

* js/class-syntax-semicolon-expected.txt: Added.
* js/class-syntax-semicolon.html: Added.
* js/script-tests/class-syntax-semicolon.js: Added.


Identifier: 162247@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@183383 268f45cc-cd09-0410-ab3c-d52691b4dbfc