AX: [ATK] REGRESSION: accessibility/canvas-fallback-content.html now crashes https://bugs.webkit.org/show_bug.cgi?id=144481
Reviewed by Chris Fleizach.

Canvas fallback content is allowed to be focusable if the canvas is displayed
and visible. Update the style when creating an accessible object for a node in
a canvas subtree and before Element::isFocusable() gets called.

No new tests. The existing, crashing test no longer crashes.

* accessibility/AXObjectCache.cpp:
(WebCore::AXObjectCache::getOrCreate):


Identifier: 162516@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@183678 268f45cc-cd09-0410-ab3c-d52691b4dbfc