[GTK] API tests crashing on debug builds due to extra unref https://bugs.webkit.org/show_bug.cgi?id=144508
Reviewed by Mario Sanchez Prada.

The problem is that we were assuming that when a new DOMWindow is
created, the DOM object cache was notified about the previous
DOMWindow being destroyed before objects for the new DOMWindow are
added to the cache. However, that's not always the case and we
only create a DOMWindowObserver for the first DOMWindow. We need
to keep a pointer to the DOMWindow being observed to clear() the
cache and create a new DOMWindowObserver when it changes in the
Frame.

Fixes crashes in several unit tests in debug builds.

* bindings/gobject/DOMObjectCache.cpp:

Identifier: 162564@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@183729 268f45cc-cd09-0410-ab3c-d52691b4dbfc